added .md so it renders as markdown
the file has markdown in it but no file type and was rendering as text way way off the right margin, etc.  I just added .md to the file name which lets GitHub know to render the file as markdown.  Now the section titles with hash tags will show up as headers of the appropriate size too.